💾

+++ scripts/daily-data-backup.sh
@@ -6,10 +6,12 @@ set -o pipefail
 readonly SOURCE_DIR="/data"
 readonly BACKUP_DIR="/backup"
 readonly DATETIME="$(date '+%Y-%m-%d_%H:%M:%S')"
-readonly LOG_DIR="${HOME}/Logs"
+readonly LOG_DIR="/data/logs"
+readonly HOMEMANAGER_DIR="${HOME}/.config/home-manager"
+readonly DOFILES_DIR="/data/dotfiles"
 readonly EXTENSIONS_DIR="${HOME}/.local/share/gnome-shell/extensions"
 readonly EXTENSIONS_DUPE_DIR="/data/extensions"
-readonly LOG_FILE="${LOG_DIR}/data_backup.log"
+readonly LOG_FILE="${LOG_DIR}/daily-data-backup.log"
 
 # Ensure the log directory exists
 mkdir -p "$LOG_DIR"
@@ -23,15 +25,47 @@ log_message() {
 }
 
 # Log start of backup
-log_message "Duplicating gnome-extensions"
-rsync -av --delete \
-    "${EXTENSIONS_DIR}" \
-    "${EXTENSIONS_DUPE_DIR}" >> "$LOG_FILE" 2>&1
-log_message "Duplication completed"
+log_message "Starting backup for date: $DATETIME"
+
+log_message "Backup home-manager"
+rsync -av --delete --ignore-existing "$HOMEMANAGER_DIR/" "$DOFILES_DIR/home-manager/"
+
+log_message "Backup from home to data dir"
+
+yq=/home/percygt/.nix-profile/bin/yq
+# Check if yq is installed
+if ! command -v $yq &> /dev/null; then
+    echo "yq is not installed. Please install it before running this script."
+    exit 1
+fi
+config="/data/rsync_home/config.yaml"
+# Read YAML using yq
+root_data_dir=$($yq -r '.root_data_dir' $config)
+root_home_dir=$($yq -r '.root_home_dir' $config)
+
+# Loop through items and perform rsync
+items=$($yq -c '.items[]' $config)
+while IFS= read -r items; do
+    data_dir=$(echo "$items" | $yq -r '.data_dir')
+    home_dir=$(echo "$items" | $yq -r '.home_dir')
+    skip_backup=$(echo "$items" | $yq -r '.skip_backup // 0' -)
+
+    # Skip backup if skip_backup is set to 1
+    if [ "$skip_backup" -eq 1 ]; then
+        echo "Skipping backup for $home_dir (skip_backup is set to 1)"
+        continue
+    fi
+    echo "from $home_dir to $data_dir"
+    # Run rsync command
+    rsync -av --delete \
+    --exclude-from=<(find $root_data_dir/$data_dir -type l -exec sh -c 'readlink -f "$0" | grep -q "/nix" && echo "$0"' {} \;) \
+    "$root_home_dir/$home_dir/" "$root_data_dir/$data_dir/"
+
+done <<< "$items"
 
 
 # Log start of backup
-log_message "Starting backup for date: $DATETIME"
+log_message "Backup from data to backup dir"
 # Create daily backup
 DAILY_DIR="${BACKUP_DIR}/daily"
 DAILY_PATH="${BACKUP_DIR}/daily/${DATETIME}"
